[FIT] changing directory
