--- App.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Form from "./Components/Form";
-import "./App.css";
-
-const App = () => {
-  return (
-    <div>
-      <Form />
-    </div>
-  );
-};
-
-export default App;
