--- Components/Form.js
@@ -1,12 +1,0 @@
-import React from "react";
-import RegisterForm from "./RegisterForm";
-
-const Form = () => {
-  return (
-    <div>
-      <RegisterForm />
-    </div>
-  );
-};
-
-export default Form;
