+++ src/App.css
@@ -1,38 +1,82 @@
-.App {
-  text-align: center;
+* {
+  margin: 0;
+  padding: 0;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #05668D;
+.container {
+  background: linear-gradient(135deg, white);
+  width: 100%;
   min-height: 100vh;
   display: flex;
-  flex-direction: column;
-  align-items: center;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
+  align-items: center;
+}
+.app-wrapper {
+  background-color: #ddd;
+  min-width: 1280px;
+  min-height: 610px;
+  padding: 30px;
+  box-sizing: border-box;
+  border-radius: 5px;
+}
+
+.title {
+  color: black;
+  text-align: center;
+  margin: 80px 0px 40px 0px;
+}
+.name,
+.email,
+.password {
+  margin: 20px;
+  color: black;
+  font-size: 16px;
+}
+
+.password {
+  margin-bottom: 40px;
+}
+.input {
+  height: 32px;
+  font-size: 20px;
+  outline: none;
+  width: 100%;
+  border-top-style: none;
+  border-left-style: none;
+  border-right-style: none;
+  border-bottom-style: 2px solid black;
+  margin-bottom: 10px;
+}
+
+.error {
+  color: red;
+}
+
+.submit {
+  display: flex;
+  text-transform: uppercase;
+  font-size: 20px;
+  border-radius: 25px;
   color: white;
+  padding: 10px 60px;
+  background: linear-gradient(to right, #303a99, #20a4f3);
+  box-shadow: rgba(0,0,0,0.3) 0px 19px 38px,
+  rgba(0,0,0,0.22) 0px 15px 12px;
+  margin: auto;
+  outline: none;
+  border: none;
+  cursor: pointer;
+  transition: all 0.1s;  
 }
 
-.App-link {
-  color: rgb(170, 207, 241);
+button:active{
+  transform: translateY(2px);
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.form.success {
+  display: flex;
+  justify-content: center;
+  margin-top: 250px;
+  min-width: 370px;
+  color: green;
 }
