+++ src/Components/Form.js
@@ -1,0 +1,12 @@
+import React from "react";
+import RegisterForm from "./RegisterForm";
+
+const Form = () => {
+  return (
+    <div>
+      <RegisterForm />
+    </div>
+  );
+};
+
+export default Form;
